+++ version_info.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+"""
+版本信息模块, 供 GUI 和 CLI 独立入口共用。
+
+Author: GuoHowe
+E-Mail: 844396800@qq.com
+Website: www.GuoHowe.com
+"""
+
+import datetime
+
+TOOL_VERSION = "1.2.1"
+TOOL_VERSION_DATE = datetime.datetime(2026, 6, 22, 0, 0, 0).strftime('%Y-%m-%d %H:%M:%S')